+++ resources/leiningen/new/luminus/reitit/src/home.clj
@@ -5,7 +5,18 @@
    [clojure.java.io :as io]
    [<<project-ns>>.middleware :as middleware]
    [ring.util.response]
-   [ring.util.http-response :as response]))
+   [ring.util.http-response :as response]<% if async %>
+   [promesa.core :as p]<% endif %>))
+<% if async %>
+(defn expensive [_]
+  (p/delay 5000 {:status 200
+                  :headers {"content-type" "text/plain"}
+                  :body "Imagine this was a long running request.
+
+But is uses promesa/CompletableFuture under the covers, so it does not tie
+up a thread - it is async."}))
+<% endif %>
+
 <% if cljs  %>
 (defn home-page [request]
   (layout/render request "home.html"))
@@ -13,9 +24,12 @@
 (defn home-routes []
   [<% if war %>"/<<name>>"<% else %>""<% endif %>
    {:middleware [middleware/wrap-csrf
-                 middleware/wrap-formats]}
+                 middleware/wrap-formats<% if async %>
+                 ;; wrap-as-async must be last
+                 middleware/wrap-as-async<% endif %>]}
    ["/" {:get home-page}]<% if graphql %>
-   ["/graphiql" {:get (fn [request] (layout/render request "graphiql.html"))}]<% endif %><% if expanded %>
+   ["/graphiql" {:get (fn [request] (layout/render request "graphiql.html"))}]<% endif %><% if async %>
+   ["/expensive" {:get expensive}]<% endif %><% if expanded %>
    ["/docs" {:get (fn [_]
                     (-> (response/ok (-> "docs/docs.md" io/resource slurp))
                         (response/header "Content-Type" "text/plain; charset=utf-8")))}]<% endif %>])
@@ -29,8 +43,11 @@
 (defn home-routes []
   [<% if war %> "/<<name>>" <% else %> "" <% endif %>
    {:middleware [middleware/wrap-csrf
-                 middleware/wrap-formats]}
-   ["/" {:get home-page}]<% if graphql %>
+                 middleware/wrap-formats
+                 <% if async %>;; wrap-as-async must be last
+                 middleware/wrap-as-async<% endif %>]}
+   ["/" {:get home-page}]<% if async %>
+   ["/expensive" {:get expensive}]<% endif %><% if graphql %>
    ["/graphiql" {:get (fn [request]
                         (layout/render request "graphiql.html"))}]<% endif %><% if expanded %>
    ["/about" {:get about-page}]<% endif %>])