+++ src/leiningen/new/undertow.clj
@@ -9,5 +9,5 @@
            :undertow-based true
            :server "undertow")
          (update :dependencies conj ['org.webjars/webjars-locator-jboss-vfs "0.1.0"])
-         (append-options :http-server-dependencies [['luminus-undertow "0.1.15"]]))]
+         (append-options :http-server-dependencies [['luminus-undertow "0.1.16"]]))]
     state))